refactor(core): move GameSystem interface
