+++ core/src/main/java/net/lapidist/colony/mod/GameServer.java
@@ -3,17 +3,17 @@ package net.lapidist.colony.mod;
 /**
  * Minimal server API exposed to mods. This placeholder avoids a build-time
  * dependency on the server module while still allowing registration of
- * {@link net.lapidist.colony.server.GameSystem} instances.
+ * {@link GameSystem} instances.
  */
 public interface GameServer {
 
     /**
-     * Register a {@link net.lapidist.colony.server.GameSystem} to start and stop
+     * Register a {@link GameSystem} to start and stop
      * with the server lifecycle.
      *
      * @param system system to manage
      */
-    default void registerSystem(net.lapidist.colony.server.GameSystem system) {
+    default void registerSystem(GameSystem system) {
         // no-op for non-server implementations
     }
 }
