--- core/src/main/java/net/lapidist/colony/server/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Interfaces shared between server and mods.
- */
-package net.lapidist.colony.server;
